+++ modules/AWSPowerShell/Cmdlets/EMRServerlessWebService/AWS.Tools.EMRServerlessWebService.Aliases.psm1
@@ -1,0 +1,3 @@
+﻿# EMRServerlessWebService
+
+Export-ModuleMember -Alias *